Issue3 WIP 2
- Extracted <PersonCard />
